+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/MindcodeCompiler.java
@@ -37,6 +37,7 @@ import info.teksol.mc.mindcode.compiler.postprocess.LogicInstructionLabelResolve
 import info.teksol.mc.mindcode.compiler.postprocess.LogicInstructionPrinter;
 import info.teksol.mc.mindcode.compiler.preprocess.DirectivePreprocessor;
 import info.teksol.mc.mindcode.compiler.preprocess.PreprocessorContext;
+import info.teksol.mc.mindcode.logic.arguments.LogicArgument;
 import info.teksol.mc.mindcode.logic.arguments.LogicVariable;
 import info.teksol.mc.mindcode.logic.instructions.InstructionProcessor;
 import info.teksol.mc.mindcode.logic.instructions.InstructionProcessorFactory;
@@ -71,7 +72,7 @@ public class MindcodeCompiler extends AbstractMessageEmitter implements AstBuild
     private final Map<InputFile, @Nullable AstIdentifier> processors = new HashMap<>();
     private final Map<InputFile, AstModule> modules = new HashMap<>();
     private final List<AstRequire> requirements = new ArrayList<>();
-    private final List<LogicVariable> remoteVariables = new ArrayList<>();
+    private final List<LogicArgument> remoteVariables = new ArrayList<>();
     private final ReturnStack returnStack;
     private final StackTracker stackTracker;
     private @Nullable AstProgram astProgram;