+++ compiler/src/test/java/info/teksol/mc/mindcode/DocValidatorTest.java
@@ -1,7 +1,6 @@
 package info.teksol.mc.mindcode;
 
 import info.teksol.mc.common.InputFiles;
-import info.teksol.mc.common.SourcePosition;
 import info.teksol.mc.messages.ListMessageLogger;
 import info.teksol.mc.messages.MindcodeMessage;
 import info.teksol.mc.mindcode.compiler.MindcodeCompiler;
@@ -11,6 +10,7 @@ import info.teksol.mc.util.CollectionUtils;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
 import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.function.Executable;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
@@ -19,10 +19,10 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 @NullMarked
 @Order(6)
@@ -53,7 +53,7 @@ public class DocValidatorTest {
     }
 
     private void validateFile(File file) throws IOException {
-        List<String> messages = new ArrayList<>();
+        List<Executable> assertions = new ArrayList<>();
         List<String> lines = Files.readAllLines(file.toPath());
         int start = 0;
 
@@ -63,7 +63,7 @@ public class DocValidatorTest {
 
             int closingIndex = CollectionUtils.indexOf(lines, index + 1, line -> line.trim().equals("```"));
             if (closingIndex < 0) {
-                messages.add("Found unterminated ```Mindcode``` block in " + file.getName() + " at line " + (index + 1));
+                assertions.add(() -> fail("Found unterminated ```Mindcode``` block in " + file.getName() + " at line " + (index + 1)));
             }
 
             String source = String.join("\n", lines.subList(index + 1, closingIndex));
@@ -74,39 +74,33 @@ public class DocValidatorTest {
                     ? String.join("\n", lines.subList(nextBlock + 1, closingNextBlock))
                     : null;
 
-            String message = validateCode(file, index + 1, source, mlog);
-            if (message != null) messages.add(message);
+            validateCode(assertions, file, index + 1, source, mlog);
 
             start = closingIndex + 1;
         }
 
-        if (!messages.isEmpty()) {
-            Assertions.fail("Validation failed:\n" + String.join("\n", messages));
-        }
+        assertAll(assertions);
     }
 
-    private @Nullable String validateCode(File file, int line, String source, @Nullable String mlog) {
+    private void validateCode(List<Executable> assertions, File file, int line, String source, @Nullable String mlog) {
         ListMessageLogger messageConsumer = new ListMessageLogger();
         InputFiles inputFiles = InputFiles.fromSource(source);
         CompilerProfile profile = createCompilerProfile();
         MindcodeCompiler compiler = new MindcodeCompiler(messageConsumer, profile, inputFiles);
         compiler.compile();
 
-        List<MindcodeMessage> messages = messageConsumer.getMessages().stream()
+        String message = messageConsumer.getMessages().stream()
                 .filter(MindcodeMessage::isErrorOrWarning)
                 .filter(m -> !m.message().matches("Variable '.*' is not used\\."))
                 .filter(m -> !m.message().matches("Variable '.*' is not initialized\\."))
-                .toList();
+                .map(m -> m.formatMessage(sp -> sp.offsetLine(line).formatForIde()))
+                .collect(Collectors.joining("\n"));
 
-        if (!messages.isEmpty()) {
-            messages.stream().map(m -> m.formatMessage(SourcePosition::formatForIde)).forEach(System.out::println);
-            return "Found errors or warnings in " + file.getName() + " at line " + line;
+        assertions.add(() -> assertTrue(message.isEmpty(), "Found errors or warnings in " + file.getName() + " at line " + line + ":\n" + message));
+
+        if (mlog != null) {
+            assertions.add(() -> assertEquals(mlog, compiler.getOutput().trim(),
+                    "Compiler output doesn't match mlog block in " + file.getName() + " at line " + line));
         }
-
-        if (mlog != null && !mlog.equals(compiler.getOutput().trim())) {
-            return "Compiler output doesn't match mlog block in " + file.getName() + " at line " + line;
-        }
-
-        return null;
     }
 }